:recycle: minor amends
