+++ src/components/banner-item/index.ts
@@ -1,1 +1,2 @@
 export * from './banner-item';
+export * from './type';
