+++ src/components/banner-item/type.ts
@@ -1,0 +1,14 @@
+export interface IBannerItemProps {
+  /**
+   *
+   */
+  banner : string;
+  /**
+   *
+   */
+  description : string;
+  /**
+   *
+   */
+  label : string;
+}
