+++ src/components/drawer-content/index.ts
@@ -1,1 +1,2 @@
 export * from './drawer-content';
+export * from './type';
