+++ src/components/drawer-content/type.ts
@@ -1,0 +1,19 @@
+export interface IDrawerContentProps {
+  /**
+   *
+   */
+  banner : string;
+  /**
+   *
+   */
+  descriptors : Record<string, unknown>;
+  /**
+   *
+   */
+  navigation : {
+    /**
+     *
+     */
+    navigate : (s : string) => void;
+  };
+}
